Added scene and event handlers.
I implemented the basic part of the GUI. You can run Game.java to start
the game. Once it starts, you will see the Main Screen. If you press any
of the buttons, a line will be printed on the console. This is for testing
purposes only. I will work on the rest.